Separates jokes into separate files
